Move AbstractIconLoader to icon/impl
